src/main/webapp/app/entities/h-1-b/route/h-1-b-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IH1B, H1B } from '../h-1-b.model';
import { H1BService } from '../service/h-1-b.service';

import { H1BRoutingResolveService } from './h-1-b-routing-resolve.service';

describe('H1B routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: H1BRoutingResolveService;
  let service: H1BService;
  let resultH1B: IH1B | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(H1BRoutingResolveService);
    service = TestBed.inject(H1BService);
    resultH1B = undefined;
  });

  describe('resolve', () => {
    it('should return IH1B returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultH1B = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultH1B).toEqual({ id: 123 });
    });

    it('should return new IH1B if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultH1B = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultH1B).toEqual(new H1B());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as H1B })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultH1B = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultH1B).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
